fix problem with too big values given in the configuration that are casted to a smaller type and strange things happen
git-svn-id: https://svn.php.net/repository/pecl/mysqlnd_ms/trunk@326429 c90b9560-bf6c-de11-be94-00142212c4b1

+++ mysqlnd_ms.c
@@ -417,7 +417,7 @@ mysqlnd_ms_connect_to_host(MYSQLND_CONN_DATA * proxy_conn, MYSQLND_CONN_DATA * c
 		char * pass_to_use = NULL;
 		char * db_to_use = NULL;
 		char * host = NULL;
-		int port, flags;
+		int64_t port, flags;
 
 		char * current_subsection_name = NULL;
 		size_t current_subsection_name_len = 0;
